requirements.txt
apache-airflow
apache-airflow-providers-amazon
pyspark
boto3
pandas
snowflake-connector-python